Try "letterbox" phrasing for embed toggle
This PR tries to rephrase the button that disables aspect-ratio based video responsiveness for the embed block.

Basically most themes _do_ apply some responsiveness to video embeds, and as such the previous "scale" phrasing could be slightly misleading. But what the feature really does, is leverage the aspect ratio of the video to ensure it's never letterboxed, i.e. has black bars on either side.

I'm not sure the phrasing here is exactly right as-is — in fact using the term letterbox seemed like a better idea in theory than in this branch — but pushing here for feedback regardless, to see if a better phrasing can emerge.